+++ atom.h
@@ -13,7 +13,7 @@ namespace simul
 		typedef Atom self_type;
 		/*TODO zamienić to poniżej na map o indeksach atom.id i kluczach *atom */
 		typedef std::set<Atom *> bond_type;
-		typedef gmtl::Vec3f position_type;
+		typedef gmtl::Vec3d position_type;
 
 	private:
 		bond_type bonds;
@@ -22,7 +22,7 @@ namespace simul
 		int id; //TODO przerobić na static z autoinkrementacja
 		position_type r;
 
-		Atom(float x, float y, float z);
+		Atom(double x, double y, double z);
 		void addBond(Atom * j, bool reverse = true);
 		int removeBond(Atom * j, bool reverse = true);
